+++ src/providers/google/ai.studio.ts
@@ -11,7 +11,7 @@ import type { EnvOverrides } from '../../types/env';
 import { maybeLoadFromExternalFile, renderVarsInObject } from '../../util';
 import { getNunjucksEngine } from '../../util/templates';
 import { parseChatPrompt, REQUEST_TIMEOUT_MS } from '../shared';
-import { CHAT_MODELS } from './googleShared';
+import { CHAT_MODELS } from './shared';
 import { maybeCoerceToGeminiFormat } from './util';
 
 const DEFAULT_API_HOST = 'generativelanguage.googleapis.com';
