+++ src/providers/google/shared.ts
