+++ src/providers/registry.ts
@@ -32,8 +32,8 @@ import { DatabricksMosaicAiChatCompletionProvider } from './databricks';
 import { EchoProvider } from './echo';
 import { FalImageGenerationProvider } from './fal';
 import { GolangProvider } from './golangCompletion';
-import { GoogleChatProvider } from './google/google';
-import { GoogleMMLiveProvider } from './google/googleLive';
+import { GoogleChatProvider } from './google/ai.studio';
+import { GoogleMMLiveProvider } from './google/live';
 import { VertexChatProvider, VertexEmbeddingProvider } from './google/vertex';
 import { GroqProvider } from './groq';
 import { HttpProvider } from './http';
