+++ test/providers/google/ai.studio.test.ts
@@ -1,5 +1,5 @@
 import * as cache from '../../../src/cache';
-import { GoogleChatProvider } from '../../../src/providers/google/google';
+import { GoogleChatProvider } from '../../../src/providers/google/ai.studio';
 import * as vertexUtil from '../../../src/providers/google/util';
 import * as templates from '../../../src/util/templates';
 
@@ -7,7 +7,7 @@ jest.mock('../../../src/cache', () => ({
   fetchWithCache: jest.fn(),
 }));
 
-jest.mock('../../../src/providers/google/vertexUtil', () => ({
+jest.mock('../../../src/providers/google/util', () => ({
   maybeCoerceToGeminiFormat: jest.fn(),
 }));
